core: move cache syscall
Moves cache syscall code to arch/arm/tee
