+++ core/arch/arm/kernel/tee_ta_manager.c
@@ -661,67 +661,6 @@ TEE_Result tee_ta_get_client_id(TEE_Identity *id)
 }
 
 /*
- * tee_uta_cache_operation - dynamic cache clean/inval request from a TA
- * It follows ARM recommendation:
- *     http://infocenter.arm.com/help/index.jsp?topic=/com.arm.doc.ddi0246d/Beicdhde.html
- * Note that this implementation assumes dsb operations are part of
- * cache_maintenance_l1(), and L2 cache sync are part of
- * cache_maintenance_l2()
- */
-#ifdef CFG_CACHE_API
-TEE_Result tee_uta_cache_operation(struct tee_ta_session *sess,
-				   enum utee_cache_operation op,
-				   void *va, size_t len)
-{
-	TEE_Result ret;
-	paddr_t pa = 0;
-	struct user_ta_ctx *utc;
-
-	if ((sess->ctx->flags & TA_FLAG_CACHE_MAINTENANCE) == 0)
-		return TEE_ERROR_NOT_SUPPORTED;
-
-	utc = to_user_ta_ctx(sess->ctx);
-	ret = tee_mmu_check_access_rights(utc,
-			TEE_MEMORY_ACCESS_WRITE, (tee_uaddr_t)va, len);
-	if (ret != TEE_SUCCESS)
-		return TEE_ERROR_ACCESS_DENIED;
-
-	ret = tee_mmu_user_va2pa(utc, va, &pa);
-	if (ret != TEE_SUCCESS)
-		return TEE_ERROR_ACCESS_DENIED;
-
-	switch (op) {
-	case TEE_CACHEFLUSH:
-		/* Clean L1, Flush L2, Flush L1 */
-		ret = cache_maintenance_l1(DCACHE_AREA_CLEAN, va, len);
-		if (ret != TEE_SUCCESS)
-			return ret;
-		ret = cache_maintenance_l2(L2CACHE_AREA_CLEAN_INV, pa, len);
-		if (ret != TEE_SUCCESS)
-			return ret;
-		return cache_maintenance_l1(DCACHE_AREA_CLEAN_INV, va, len);
-
-	case TEE_CACHECLEAN:
-		/* Clean L1, Clean L2 */
-		ret = cache_maintenance_l1(DCACHE_AREA_CLEAN, va, len);
-		if (ret != TEE_SUCCESS)
-			return ret;
-		return cache_maintenance_l2(L2CACHE_AREA_CLEAN, pa, len);
-
-	case TEE_CACHEINVALIDATE:
-		/* Inval L2, Inval L1 */
-		ret = cache_maintenance_l2(L2CACHE_AREA_INVALIDATE, pa, len);
-		if (ret != TEE_SUCCESS)
-			return ret;
-		return cache_maintenance_l1(DCACHE_AREA_INVALIDATE, va, len);
-
-	default:
-		return TEE_ERROR_NOT_SUPPORTED;
-	}
-}
-#endif
-
-/*
  * dump_state - Display TA state as an error log.
  */
 static void dump_state(struct tee_ta_ctx *ctx)
