+++ core/arch/arm/tee/sub.mk
@@ -2,6 +2,7 @@ srcs-y += arch_tee_fs.c
 srcs-y += tee_rpmb.c
 srcs-$(CFG_ARM32_core) += arch_svc_a32.S
 srcs-$(CFG_ARM64_core) += arch_svc_a64.S
+srcs-$(CFG_CACHE_API) += svc_cache.c
 srcs-y += arch_svc.c
 srcs-y += entry_std.c
 srcs-y += entry_fast.c
