+++ core/tee/tee_svc.c
@@ -909,20 +909,3 @@ TEE_Result syscall_set_ta_time(const TEE_Time *mytime)
 
 	return tee_time_set_ta_time((const void *)&s->ctx->uuid, &t);
 }
-
-#ifdef CFG_CACHE_API
-TEE_Result syscall_cache_operation(void *va, size_t len, unsigned long op)
-{
-	TEE_Result res;
-	struct tee_ta_session *s = NULL;
-
-	res = tee_ta_get_current_session(&s);
-	if (res != TEE_SUCCESS)
-		return res;
-
-	if ((s->ctx->flags & TA_FLAG_CACHE_MAINTENANCE) == 0)
-		return TEE_ERROR_NOT_SUPPORTED;
-
-	return tee_uta_cache_operation(s, op, va, len);
-}
-#endif
